🎨 refactor: finaliza refatorção header
