+++ src/components/Header/Header.css
@@ -1,6 +1,6 @@
-.cover-page {
+.header-container {
   max-width: 600px;
-  margin: 10px auto 0;
+  margin: 0 auto 0;
 }
 
 .cover-page img {
