+++ src/components/Main/Main.css
@@ -1,0 +1,4 @@
+.main-container {
+  max-width: 600px;
+  margin: 0 auto 0;
+}
