+++ src/components/Navbar/Navbar.css
@@ -1,15 +1,10 @@
-.main-home {
-  max-width: 600px;
-  margin: 0 auto 0;
-}
-
-.header-main {
+.nav-container {
   text-align: center;
   display: flex;
   justify-content: space-around;
 }
 
-.h3-header-main {
+.h3-nav-main {
   font-size: 1.2rem;
   font-weight: 700;
   padding: 0 1rem;
@@ -29,19 +24,6 @@
 .active {
   border-bottom: 3px solid var(--pink);
   color: white;
-}
-
-.links-main {
-  text-decoration: none;
-  color: grey;
-}
-
-.footer-container {
-  text-align: center;
-  width: 100%;
-  margin-top: 2rem;
-  font-size: 0.8rem;
-  color: grey;
 }
 
 @media screen and (max-width: 407px) {
