+++ src/pages/Feed.jsx
@@ -4,7 +4,7 @@ import '../CSS/Feed.css';
 
 import dataFeed from '../api/feed';
 
-import Main from '../components/Main';
+import Main from '../components/Main/Main';
 import profile from '../images/profile.png';
 
 export default function Feed() {
